Remove Walter
I decided the simplest way to alleviate the reviewer's concern was to simply remove the discussion of Walter, as it is a motivating example only. So I rewrote with 2 archetypes instead of 3.

It would also be possible to keep Walter in, but position disarmament as the second example, rather than the first. However, I think the reviewer is right that the destruction of weapons is not a major point of contention usually. The key issue is the rebels giving up their weapons -- the government (usually) doesn't really care if they are confiscated vs. destroyed.  So it isn't a great archetype.